requirements.txt
boto3==1.26.118
deepface==0.0.79
Django==4.2
django_environ==0.10.0
gpsphoto==2.2.3
numpy==1.24.2
Pillow==9.5.0
pillow_heif==0.10.1
Requests==2.29.0
scipy==1.10.1
tqdm==4.65.0
transformers==4.28.1
exifread==3.0.0
piexif==1.1.3
tensorflow==2.10.1
torch==2.0.0
